move fns used only by macros into their own module
