+++ crates/pico/src/lib.rs
@@ -1,10 +1,3 @@
-use std::{
-    any::Any,
-    hash::{DefaultHasher, Hash, Hasher},
-};
-
-use crate::index::Index;
-
 mod arena;
 mod database;
 mod dependency;
@@ -13,6 +6,7 @@ mod dyn_eq;
 mod epoch;
 mod generation;
 mod index;
+pub mod macro_fns;
 mod memo;
 mod source;
 mod u64_types;
@@ -22,31 +16,3 @@ pub use derived_node::*;
 pub use memo::*;
 pub use source::*;
 pub use u64_types::*;
-
-pub fn intern_param<T: Hash + Clone + 'static>(db: &Database, param: T) -> ParamId {
-    let param_id = hash(&param).into();
-    if !db.contains_param(param_id) {
-        let idx = db
-            .epoch_to_generation_map
-            .get(&db.current_epoch)
-            .unwrap()
-            .insert_param(Box::new(param));
-        db.param_id_to_index
-            .insert(param_id, Index::new(db.current_epoch, idx));
-    }
-    param_id
-}
-
-pub fn get_derived_node(db: &Database, derived_node_id: DerivedNodeId) -> Option<&DerivedNode> {
-    db.get_derived_node(derived_node_id)
-}
-
-pub fn get_param(db: &Database, param_id: ParamId) -> Option<&Box<dyn Any>> {
-    db.get_param(param_id)
-}
-
-pub fn hash<T: Hash>(value: &T) -> u64 {
-    let mut s = DefaultHasher::new();
-    value.hash(&mut s);
-    s.finish()
-}
